refactor(tui): extract useRouteProbe hook
Move probe state (probeMode/probeModel/probeResults) and the route+test
state machine (App.tsx submit IIFE) into hooks/useRouteProbe.ts. App.tsx
keyboard handler delegates via verb methods: startInput, typeChar,
backspace, submit, cancel, enterFromDone.

Discriminated-union state (idle/input/running/done) is exposed alongside
the legacy probeMode/probeModel/probeResults atoms for the render
components. Plain useState under the hood — no useReducer.

INTENTIONAL: cancel() does NOT abort the in-flight async test loop. The
loop keeps writing to setProbeResults / setProbeMode after the user
cancels, so results can pop back into "done" — same as baseline. Adding
an AbortController would change observable behavior. See the type
comment on ProbeState.

App.tsx 959 -> 870 LOC.
